literary_composer/pyproject.toml
[project]
name = "langchain-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.12.0",
    "certifi>=2025.11.12",
    "isort>=7.0.0",
    "langchain>=1.2.0",
    "langchain-google-genai>=4.1.2",
    "langchain-groq>=1.1.1",
    "langchain-ollama>=1.0.1",
    "ollama>=0.6.1",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.2",
]
